rewrite to Webpack and ES6
The entire project has been casually rewritten to ES6 and Webpack, in a
single, lean, absolutely not bloated commit.
Also, it has been modularized a bit.
