+++ package.json
@@ -1,27 +1,45 @@
 {
-    "name"         : "map-generator",
-    "version"      : "0.1.0",
-    "description"  : "JS random dungeon map generator",
-    "author"       : "Robert Kardos",
-    "email"        : "kardos.robert.kh@gmail.com",
-    "repository"   : {
-        "type" : "git",
-        "url"  : "https://github.com/Kedveshentes/map-generator"
-    },
-    "keywords"     : [
-        "random",
-        "map",
-        "dungeon",
-        "combobreaker",
-        "generator"
-    ],
-    "dependencies" : {
-        "gl-matrix"  : "~2.3.0",
-        "requirejs"  : "~2.1.20",
-        "underscore" : "*",
-        "three"      : "*",
-        "mime"       : "~1.3.4",
-        "fps"        : "*"
-    },
-    "license"      : "MIT"
+  "name": "map-generator",
+  "version": "0.1.0",
+  "description": "JS random dungeon map generator",
+  "author": "Robert Kardos",
+  "email": "kardos.robert.kh@gmail.com",
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/Kedveshentes/map-generator.git"
+  },
+  "keywords": [
+    "random",
+    "map",
+    "dungeon",
+    "combobreaker",
+    "generator"
+  ],
+  "dependencies": {
+    "babel-polyfill": "^6.9.1",
+    "babel-runtime": "^6.9.2",
+    "fps": "0.0.3",
+    "gl-matrix": "~2.3.0",
+    "three": "^0.78.0",
+    "underscore": "^1.8.3"
+  },
+  "license": "MIT",
+  "devDependencies": {
+    "babel": "^6.5.2",
+    "babel-core": "^6.9.1",
+    "babel-loader": "^6.2.4",
+    "babel-plugin-transform-runtime": "^6.9.0",
+    "babel-preset-es2015": "^6.9.0",
+    "webpack": "^1.13.1",
+    "webpack-dev-server": "^1.14.1"
+  },
+  "scripts": {
+    "start": "webpack",
+    "dev": "webpack-dev-server --inline"
+  },
+  "bugs": {
+    "url": "https://github.com/Kedveshentes/map-generator/issues"
+  },
+  "homepage": "https://github.com/Kedveshentes/map-generator#readme",
+  "main": "server.js"
 }
