--- public/index.html
@@ -1,18 +1,0 @@
-<!DOCTYPE html>
-<html>
-<head>
-	<link rel="stylesheet" href="style.css">
-	<title>Map generator</title>
-</head>
-<body>
-	<canvas id="canvas" width="1" height="1"></canvas>
-
-	<!-- <script src="node_modules/three/three.min.js"></script>
-	<script src="node_modules/underscore/underscore-min.js"></script>
-	<script src="node_modules/gl-matrix/dist/gl-matrix-min.js"></script> -->
-	<script data-main="js/app.js" src="require.js"></script>
-	<!-- <script src="wat.js"></script> -->
-
-
-</body>
-</html>
