--- public/js/wallFactory.js
@@ -1,4 +1,0 @@
-define(['three', 'util'],
-function (THREE, util) {
-
-});
