--- public/js/wat.js
@@ -1,1 +1,0 @@
-export const kalacs = 'KALACS';
