--- public/wat.js
@@ -1,7 +1,0 @@
-window.main = function () {
-    window.requestAnimationFrame( main );
-    console.log('asd');
-  // Whatever your main loop needs to do.
-};
-
-main();
