+++ src/index.html
@@ -1,0 +1,11 @@
+<!DOCTYPE html>
+<html>
+<head>
+	<title>Map generator</title>
+	<link rel="stylesheet" href="style.css">
+</head>
+<body>
+	<canvas id="canvas" width="1" height="1"></canvas>
+	<script src="../public/app.bundle.js" charset="utf-8"></script>
+</body>
+</html>
